fix(EMA): avoid adding -d flag to plans not using them
The question one might have is whether you couldn't just ignore
the detectors on the plan parsing code.

The reason we can't do that here is because the `mov` plan takes
motors and positions as positional arguments. That means that having
the `-d` flag eats every entry, and we become sad.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>