feat(EMA): create a preprocessor for injecting readings into a plan
This builds off the previous work with the input processor, creating a
whole system for specifying plan parameters in a reasonably automatic
manner.

The preprocessor work will possibly move to a more general place,
potentially sophys-common, in the near future, as we must make this
system work in a remote setting too. The code core is reasonably
mode-agnostic as it is, so it shouldn't be much of a chore to do.

I still don't know whether we ought to use bps.read or bps.trigger_and_read
though, haven't tested this in a real setting yet.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>