feat(EMA): support flexible k-space range parameters in escan
This adds support for specifying k-space ranges as a pair of
end position and step, with the start position being taken
from the previous range (either normal or k-space).

This simply provides the necessary information to support such
processing. Building the actual trajectory will be left to the
plan itself.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>